+++ packages/web-core/src/index.ts
@@ -42,6 +42,12 @@ import {
   SessionContinuation,
   SessionContinuationProps,
 } from './entities/continuation/session_continuation.entity';
+import {
+  ErrorReconnectionType,
+  ErrorResourceType,
+  ErrorType,
+  InworldError,
+} from './entities/error.entity';
 import { DislikeType, Feedback } from './entities/feedback.entity';
 import {
   AdditionalPhonemeInfo,
@@ -91,6 +97,9 @@ export {
   EmotionEvent,
   EmotionStrength,
   EmotionStrengthCode,
+  ErrorReconnectionType,
+  ErrorResourceType,
+  ErrorType,
   Extension,
   Feedback,
   FeedbackDislikeProps,
@@ -105,6 +114,7 @@ export {
   HistoryItemTriggerEvent,
   InworldClient,
   InworldConnectionService,
+  InworldError,
   InworldPacket,
   InworldTriggers,
   MicrophoneMode,